Add a check to unfeasible analyzer : move constraint slack analysis to separate source files
