+++ src/solver/infeasible-problem-analysis/unfeasible-pb-analysis.h
@@ -6,27 +6,9 @@
 
 #include "../utils/ortools_utils.h"
 #include "unfeability-diagnostic.h"
-#include "report.h"
 
 namespace Antares::Optimization
 {
-
-class ConstraintSlackDiagnostic : public InfeasibilityDiagnostic
-{
-    using InfeasibilityDiagnostic::InfeasibilityDiagnostic;
-public:
-    void run() override;
-    void printReport() override;
-    std::string title() override { return "Slack variables diagnostic"; }
-
-private:
-    void buildObjective() const;
-    void addSlackVariables();
-    operations_research::MPSolver::ResultStatus Solve() const;
-
-    std::vector<const operations_research::MPVariable*> slackVariables_;
-    const std::string constraint_name_pattern = "^AreaHydroLevel::|::hourly::|::daily::|::weekly::|^FictiveLoads::";
-};
 
 struct VariableBounds
 {
